Clean up frontend-api folder of dead code [Re: nmkataoka/adv-life#133]
Removed old getter methods that are no longer being used due to new
react-ecsal selectors.

Did some light renaming.
