+++ src/3-frontend-api/combat/index.ts
@@ -1,0 +1,1 @@
+export { setSkillTarget } from './setSkillTarget';
