+++ src/3-frontend-api/name/getName.ts
@@ -1,6 +1,6 @@
 import { EntityManager } from '0-engine';
 import { NameCmpt } from '1-game-code/ncomponents';
 
-export const getNameCmpt = (e: number) => (eMgr: EntityManager): NameCmpt => {
+export const getName = (e: number) => (eMgr: EntityManager): NameCmpt => {
   return eMgr.getCmpt(NameCmpt, e);
 };
