+++ src/3-frontend-api/name/index.ts
@@ -1,0 +1,1 @@
+export { getName } from './getName';
