+++ src/3-frontend-api/town/getTownLocation.ts
@@ -1,12 +1,12 @@
 import { EntityManager } from '0-engine';
 import { getInventory } from '3-frontend-api/inventory';
-import { getNameCmpt } from '../getName';
+import { getName } from '../name/getName';
 import { TownLocationInfo } from './TownLocationInfo';
 
 export const getTownLocation = (townLocationId: number) => (
   eMgr: EntityManager,
 ): TownLocationInfo => {
-  const { name = 'Unnamed' } = getNameCmpt(townLocationId)(eMgr);
+  const { name = 'Unnamed' } = getName(townLocationId)(eMgr);
   const inventory = getInventory(townLocationId)(eMgr);
 
   return {
